``` feat: Refactor routes and handlers
This commit introduces a refactoring of the application's route handling and logic by moving route definitions into dedicated handler files.

The commit includes the following changes:

- Removed the route files
- Created handler files
- Fixed all the imports/exports

This change improves the overall organization and maintainability of the codebase.
```
