+++ package.json
@@ -24,20 +24,12 @@
     "lint": "biome check --formatter-enabled=true --linter-enabled=true --organize-imports-enabled=true --fix src"
   },
   "dependencies": {
-    "@elysiajs/cors": "^1.3.3",
-    "@elysiajs/html": "^1.3.0",
-    "@elysiajs/server-timing": "^1.3.0",
-    "@elysiajs/static": "^1.3.0",
-    "@elysiajs/swagger": "^1.3.0",
     "chalk": "^5.4.1",
     "date-fns": "^4.1.0",
     "docker-compose": "^1.2.0",
     "dockerode": "^4.0.7",
-    "elysia": "latest",
-    "elysia-remote-dts": "^1.0.3",
     "js-yaml": "^4.1.0",
     "knip": "latest",
-    "logestic": "^1.2.4",
     "split2": "^4.2.0",
     "winston": "^3.17.0",
     "yaml": "^2.8.0"
