+++ src/handlers/database.ts
@@ -1,0 +1,13 @@
+import { dbFunctions } from "~/core/database";
+
+class databaseHandler {
+  async getContainers() {
+    return dbFunctions.getContainerStats();
+  }
+
+  async getHosts() {
+    return dbFunctions.getHostStats();
+  }
+}
+
+export const DatabaseHandler = new databaseHandler();
