--- src/routes/logs.ts
@@ -1,261 +1,0 @@
-import { Elysia } from "elysia";
-
-import { dbFunctions } from "~/core/database";
-import { logger } from "~/core/utils/logger";
-
-export const backendLogs = new Elysia({ prefix: "/logs" })
-	.get(
-		"",
-		async ({ set }) => {
-			try {
-				const logs = dbFunctions.getAllLogs();
-				//
-				logger.debug("Retrieved all logs");
-				return logs;
-			} catch (error) {
-				set.status = 500;
-				logger.error("Failed to retrieve logs,", error);
-				return { error: "Failed to retrieve logs" };
-			}
-		},
-		{
-			detail: {
-				tags: ["Management"],
-				description:
-					"Retrieves complete application log history from persistent storage",
-				responses: {
-					"200": {
-						description: "Successfully retrieved logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "array",
-									items: {
-										type: "object",
-										properties: {
-											id: {
-												type: "number",
-												example: 1,
-											},
-											level: {
-												type: "string",
-												example: "info",
-											},
-											message: {
-												type: "string",
-												example: "Application started",
-											},
-											timestamp: {
-												type: "string",
-												example: "2024-03-20T12:00:00Z",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					"500": {
-						description: "Error retrieving logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										error: {
-											type: "string",
-											example: "Failed to retrieve logs",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-
-	.get(
-		"/:level",
-		async ({ params: { level }, set }) => {
-			try {
-				const logs = dbFunctions.getLogsByLevel(level);
-
-				logger.debug(`Retrieved logs (level: ${level})`);
-				return logs;
-			} catch (error) {
-				set.status = 500;
-				logger.error("Failed to retrieve logs");
-				return { error: "Failed to retrieve logs" };
-			}
-		},
-		{
-			detail: {
-				tags: ["Management"],
-				description:
-					"Filters logs by severity level (debug, info, warn, error, fatal)",
-				responses: {
-					"200": {
-						description: "Successfully retrieved logs by level",
-						content: {
-							"application/json": {
-								schema: {
-									type: "array",
-									items: {
-										type: "object",
-										properties: {
-											id: {
-												type: "number",
-												example: 1,
-											},
-											level: {
-												type: "string",
-												example: "info",
-											},
-											message: {
-												type: "string",
-												example: "Application started",
-											},
-											timestamp: {
-												type: "string",
-												example: "2024-03-20T12:00:00Z",
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-					"500": {
-						description: "Error retrieving logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										error: {
-											type: "string",
-											example: "Failed to retrieve logs",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-
-	.delete(
-		"/",
-		async ({ set }) => {
-			try {
-				set.status = 200;
-
-				dbFunctions.clearAllLogs();
-				return { success: true };
-			} catch (error) {
-				set.status = 500;
-				logger.error("Could not delete all logs,", error);
-				return { error: "Could not delete all logs" };
-			}
-		},
-		{
-			detail: {
-				tags: ["Management"],
-				description: "Purges all historical log records from the database",
-				responses: {
-					"200": {
-						description: "Successfully cleared all logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										success: {
-											type: "boolean",
-											example: true,
-										},
-									},
-								},
-							},
-						},
-					},
-					"500": {
-						description: "Error clearing logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										error: {
-											type: "string",
-											example: "Could not delete all logs",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-
-	.delete(
-		"/:level",
-		async ({ params: { level }, set }) => {
-			try {
-				dbFunctions.clearLogsByLevel(level);
-
-				logger.debug(`Cleared all logs with level: ${level}`);
-				return { success: true };
-			} catch (error) {
-				set.status = 500;
-				logger.error("Could not clear logs with level", level, ",", error);
-				return { error: "Failed to retrieve logs" };
-			}
-		},
-		{
-			detail: {
-				tags: ["Management"],
-				description: "Clears log entries matching specified severity level",
-				responses: {
-					"200": {
-						description: "Successfully cleared logs by level",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										success: {
-											type: "boolean",
-											example: true,
-										},
-									},
-								},
-							},
-						},
-					},
-					"500": {
-						description: "Error clearing logs",
-						content: {
-							"application/json": {
-								schema: {
-									type: "object",
-									properties: {
-										error: {
-											type: "string",
-											example: "Failed to retrieve logs",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	);
